move element implementation details to shared trait for input and text-area
