+++ crates/gpui_elements/src/editable_text/text_area_element.rs
@@ -1,10 +1,11 @@
 use crate::editable_text::{
-    InitStorage, TextAreaState,
+    InitStorage, StateBackedEditableText, TextAreaState,
     actions::{DEFAULT_INPUT_CONTEXT, EditableInputActionElement},
+    shared_element::{self, EditableTextElement},
 };
 use gpui::{
-    Element, ElementId, Entity, Hitbox, InteractiveElement, Interactivity, IntoElement,
-    SharedString, StyleRefinement, Styled, TextStyle, WeakEntity,
+    App, Bounds, Element, ElementId, Entity, Hitbox, InteractiveElement, Interactivity,
+    IntoElement, Pixels, SharedString, StyleRefinement, Styled, TextStyle, WeakEntity, Window,
 };
 use std::{cell::RefCell, rc::Rc};
 
@@ -34,6 +35,13 @@ pub struct TextAreaElement {
     placeholder: Option<SharedString>,
 }
 
+impl TextAreaElement {
+    pub fn placeholder(mut self, text: impl Into<SharedString>) -> Self {
+        self.placeholder = Some(text.into());
+        self
+    }
+}
+
 impl InteractiveElement for TextAreaElement {
     fn interactivity(&mut self) -> &mut Interactivity {
         &mut self.interactivity
@@ -53,31 +61,33 @@ impl IntoElement for TextAreaElement {
     }
 }
 
-impl EditableInputActionElement for TextAreaElement {
+impl StateBackedEditableText for TextAreaElement {
     type State = TextAreaState;
+}
+
+impl EditableInputActionElement for TextAreaElement {
     fn state_entity_rc(&self) -> &Rc<RefCell<WeakEntity<Self::State>>> {
         &self.state_entity
     }
 }
 
-pub mod element {
-    use super::*;
-
-    #[doc(hidden)]
-    pub struct LayoutState {
-        pub state: Entity<TextAreaState>,
-        pub text_style: TextStyle,
+impl EditableTextElement for TextAreaElement {
+    fn init_state(&self, cx: &mut gpui::prelude::Context<Self::State>) -> Self::State {
+        Self::State::new(self.init_storage.exec(cx), cx)
     }
 
-    #[doc(hidden)]
-    pub struct PrepaintState {
-        pub hitbox: Option<Hitbox>,
+    fn placeholder(&self) -> &Option<SharedString> {
+        &self.placeholder
+    }
+
+    fn should_wrap(&self) -> bool {
+        true
     }
 }
 
 impl Element for TextAreaElement {
-    type RequestLayoutState = element::LayoutState;
-    type PrepaintState = element::PrepaintState;
+    type RequestLayoutState = shared_element::LayoutState<TextAreaState>;
+    type PrepaintState = shared_element::PrepaintState;
 
     fn id(&self) -> Option<ElementId> {
         self.interactivity.element_id.clone()
@@ -91,74 +101,42 @@ impl Element for TextAreaElement {
         &mut self,
         global_id: Option<&gpui::GlobalElementId>,
         inspector_id: Option<&gpui::InspectorElementId>,
-        window: &mut gpui::Window,
-        cx: &mut gpui::App,
+        window: &mut Window,
+        cx: &mut App,
     ) -> (gpui::LayoutId, Self::RequestLayoutState) {
-        // Fetches or initializes the internal state of the field
-        let state = match &self.interactivity.element_id {
-            None => unimplemented!("all input elements must be assigned an id"),
-            Some(element_id) => {
-                let state = window.use_keyed_state(element_id.clone(), cx, |_window, cx| {
-                    TextAreaState::new(self.init_storage.exec(cx), cx)
-                });
-                // store a reference to the entity owned by the element for access in action handlers
-                *self.state_entity.borrow_mut() = state.downgrade();
-                state
-            }
-        };
-
-        let mut resolved_text_style = None;
-
-        let layout_id = self.interactivity.request_layout(
-            global_id,
-            inspector_id,
-            window,
-            cx,
-            |element_style, window, cx| {
-                window.with_text_style(element_style.text_style().cloned(), |window| {
-                    resolved_text_style = Some(window.text_style());
-
-                    let mut style = element_style.clone();
-                    if let gpui::Length::Auto = style.size.width {
-                        style.size.width = gpui::relative(1.).into();
-                    }
-                    if let gpui::Length::Auto = style.size.height {
-                        style.size.height = gpui::relative(1.).into();
-                    }
-                    window.request_layout(style, None, cx)
-                })
-            },
-        );
-
-        let layout_state = element::LayoutState {
-            state,
-            text_style: resolved_text_style.unwrap_or_else(|| window.text_style()),
-        };
-        (layout_id, layout_state)
+        self.shared_request_layout(global_id, inspector_id, window, cx)
     }
 
     fn prepaint(
         &mut self,
-        id: Option<&gpui::GlobalElementId>,
+        global_id: Option<&gpui::GlobalElementId>,
         inspector_id: Option<&gpui::InspectorElementId>,
-        bounds: gpui::Bounds<gpui::Pixels>,
+        bounds: Bounds<Pixels>,
         request_layout: &mut Self::RequestLayoutState,
-        window: &mut gpui::Window,
-        cx: &mut gpui::App,
+        window: &mut Window,
+        cx: &mut App,
     ) -> Self::PrepaintState {
-        todo!()
+        self.shared_prepaint(global_id, inspector_id, bounds, request_layout, window, cx)
     }
 
     fn paint(
         &mut self,
-        id: Option<&gpui::GlobalElementId>,
+        global_id: Option<&gpui::GlobalElementId>,
         inspector_id: Option<&gpui::InspectorElementId>,
-        bounds: gpui::Bounds<gpui::Pixels>,
+        bounds: Bounds<Pixels>,
         request_layout: &mut Self::RequestLayoutState,
         prepaint: &mut Self::PrepaintState,
-        window: &mut gpui::Window,
-        cx: &mut gpui::App,
+        window: &mut Window,
+        cx: &mut App,
     ) {
-        todo!()
+        self.shared_paint(
+            global_id,
+            inspector_id,
+            bounds,
+            request_layout,
+            prepaint,
+            window,
+            cx,
+        );
     }
 }
